refactor(detection): move Recognizer trait to engine, flatten layout
Two related restructures landed together.

## Recognizer + DetectionParams move back to nvisy-engine

`nvisy-core::detection` was added in #156 with the prospect that
backend crates (pattern, nlp, agent) might implement `Recognizer`
directly. That didn't materialize — every implementation today is
an engine-side adapter wrapping a backend, and nothing outside
nvisy-engine references the trait. Keeping it in core was
speculation, so it goes back where its only consumers live:

- `nvisy-core/src/detection/{mod.rs, params.rs}` deleted.
- `nvisy-core` drops `async-trait` + `validator` (only needed for
  the recognizer trait + DetectionParams validation).
- `Recognizer` trait now lives at `nvisy-engine/src/detection/
  recognizer.rs` (own file).
- `DetectionParams` at `nvisy-engine/src/detection/params.rs`.
- Internal imports `use nvisy_core::detection::Recognizer` →
  `use crate::detection::Recognizer`.

## Detection module flattened

`detection/engine/` and `detection/recognizer/` subfolders go away.
The orchestrator code merges into `detection/mod.rs`; the adapter
folders move up one level to sit alongside the trait.

Adapter folders also shorten their names: `language_model/` →
`llm/`, `named_entity/` → `ner/`. Pattern keeps its name.

Each adapter's per-call context (`LlmContext`, `NerContext`,
`PatternContext`) moves out of the adapter `mod.rs` into its own
`context.rs` next to the recognizer struct + params:

  detection/
    mod.rs                  ← DetectionEngine + Detection workflow + builder
    recognizer.rs           ← Recognizer trait
    context.rs              ← DetectionContext (the fat per-call input)
    dyn_recognizer.rs       ← DynRecognizer bridge trait
    params.rs               ← DetectionParams
    extension/              ← RebaseEntities
    llm/
      mod.rs                ← LlmRecognizer + impls
      context.rs            ← LlmContext + From<&DetectionContext>
      params.rs             ← LlmDetection
    ner/                    ← same shape (NerRecognizer / NerContext / NerDetection)
    pattern/                ← same shape (PatternRecognizer / PatternContext / PatternDetection)

Tracing target renamed from `nvisy_engine::detection::engine` →
`nvisy_engine::detection`. Adapter docstring intra-doc links
updated to point at the new flat paths. Doc warnings for
`super::*Recognizer::run` (trait method, not resolvable as an
intra-doc link) replaced with backtick-only references to the
recognizer struct.

No behavior change — pure restructure.
