+++ crates/nvisy-core/src/lib.rs
@@ -3,7 +3,6 @@
 #![doc = include_str!("../README.md")]
 
 pub mod content;
-pub mod detection;
 pub mod media;
 
 mod error;
