+++ crates/nvisy-engine/src/detection/context.rs
