+++ crates/nvisy-engine/src/detection/llm/params.rs
