+++ crates/nvisy-engine/src/detection/ner/mod.rs
@@ -17,53 +17,19 @@
 //! [`from_engine`]: NerRecognizer::from_engine
 //! [`engine`]: NerDetection::engine
 
+mod context;
 mod params;
 
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use nvisy_codec::handler::TextData;
 use nvisy_core::Result;
-use nvisy_core::detection::Recognizer;
 use nvisy_nlp::{Context as NlpContext, Engine as NlpEngine};
-use nvisy_ontology::entity::{Entities, EntityKind};
-use nvisy_ontology::primitive::LanguageTag;
-use uuid::Uuid;
+use nvisy_ontology::entity::Entities;
 
+pub use self::context::NerContext;
 pub use self::params::NerDetection;
-use crate::detection::DetectionContext;
-
-/// Per-call input to [`NerRecognizer::run`].
-#[derive(Debug, Clone)]
-pub struct NerContext {
-    /// The text to analyze.
-    pub text: TextData,
-    /// Caller-asserted language. When `Some`, NER skips per-call
-    /// language detection.
-    pub language: Option<LanguageTag>,
-    /// Restrict language detection to this subset. Ignored when
-    /// `language` is `Some`.
-    pub candidate_languages: Option<Vec<LanguageTag>>,
-    /// Entity-kind allowlist. Empty = all kinds permitted.
-    pub entities: Option<Vec<EntityKind>>,
-    /// Minimum confidence threshold in `[0.0, 1.0]`.
-    pub score_threshold: Option<f64>,
-    /// Correlation UUID propagated through the tracing span.
-    pub correlation_id: Option<Uuid>,
-}
-
-impl From<&DetectionContext> for NerContext {
-    fn from(ctx: &DetectionContext) -> Self {
-        Self {
-            text: ctx.text.clone(),
-            language: ctx.language.clone(),
-            candidate_languages: ctx.candidate_languages.clone(),
-            entities: ctx.entities.clone(),
-            score_threshold: ctx.score_threshold,
-            correlation_id: ctx.correlation_id,
-        }
-    }
-}
+use crate::detection::Recognizer;
 
 /// NER recognizer backed by [`nvisy_nlp::Engine`].
 pub struct NerRecognizer {
