+++ crates/nvisy-engine/src/detection/params.rs
