+++ crates/nvisy-engine/src/detection/pattern/params.rs
