+++ crates/nvisy-engine/src/detection/recognizer.rs
@@ -1,28 +1,11 @@
-//! [`Recognizer`] trait and shared per-call detection config.
+//! [`Recognizer`] trait — the per-recognizer abstraction the
+//! [`DetectionEngine`] dispatches against.
 //!
-//! A recognizer is the unit of entity-detection plug-in: it consumes
-//! a typed per-call context (chosen by the impl as its associated
-//! [`Recognizer::Context`]) and produces [`Entities`].
-//!
-//! The trait carries the *bare minimum* needed to dispatch:
-//! associated context type, async `run`, optional `reset` for
-//! cross-call state. Orchestration, parallel dispatch, and the
-//! wider per-call context type live in downstream crates.
-//!
-//! Each backend defines its own context shape next to its
-//! recognizer adapter. A bridge trait downstream converts a
-//! universal "fat" context into each adapter's typed slice via
-//! `From<&FatContext>`.
-//!
-//! [`Entities`]: nvisy_ontology::entity::Entities
-
-mod params;
+//! [`DetectionEngine`]: super::DetectionEngine
 
 use async_trait::async_trait;
+use nvisy_core::Result;
 use nvisy_ontology::entity::Entities;
-
-pub use self::params::DetectionParams;
-use crate::Result;
 
 /// Recognize entities given a per-recognizer context.
 ///
@@ -32,11 +15,13 @@ use crate::Result;
 /// recognizers need their own per-call config. No common context
 /// type appears in this trait.
 ///
-/// Implementations are independent — orchestrators run each on its
-/// own and merge results.
+/// Implementations are independent — the orchestrator
+/// ([`DetectionEngine`]) runs each on its own and merges results.
 ///
 /// Async because realistic impls dispatch to ONNX inference on a
 /// blocking pool or call remote services.
+///
+/// [`DetectionEngine`]: super::DetectionEngine
 #[async_trait]
 pub trait Recognizer: Send + Sync {
     /// The per-call context this recognizer consumes.
